requirements.txt
gradio>=4.0
lightrag-hku
google-generativeai
pdfplumber
supabase
python-dotenv
numpy
tenacity
asyncpg
